\name{worldCOVID}
\alias{worldCOVID}
\docType{data}
\title{World's COVID-19 number of daily new deaths
}
\description{Dataset for weibull4 package. This dataset contain the number of the World's
new daily deaths for COVID-19 to be fitted to the 4-parameters Weibull distribution
in weibull4 package.
}
\usage{data("worldCOVID")}
\format{
  A data frame with 157 observations on the following 2 variables.
  \describe{
    \item{\code{date}}{as.Date(c("2019-12-31", "2020-01-01", "2020-01-02", "2020-01-03", "2020-01-04", "2020-01-05", "2020-01-06", "2020-01-07", "2020-01-08", "2020-01-09", "2020-01-10", "2020-01-11", "2020-01-12", "2020-01-13", "2020-01-14", "2020-01-15", "2020-01-16", "2020-01-17", "2020-01-18", "2020-01-19", "2020-01-20", "2020-01-21", "2020-01-22", "2020-01-23", "2020-01-24", "2020-01-25", "2020-01-26", "2020-01-27", "2020-01-28", "2020-01-29", "2020-01-30", "2020-01-31", "2020-02-01", "2020-02-02", "2020-02-03", "2020-02-04", "2020-02-05", "2020-02-06", "2020-02-07", "2020-02-08", "2020-02-09", "2020-02-10", "2020-02-11", "2020-02-12", "2020-02-13", "2020-02-14", "2020-02-15", "2020-02-16", "2020-02-17", "2020-02-18", "2020-02-19", "2020-02-20", "2020-02-21", "2020-02-22", "2020-02-23", "2020-02-24", "2020-02-25", "2020-02-26", "2020-02-27", "2020-02-28", "2020-02-29", "2020-03-01", "2020-03-02", "2020-03-03", "2020-03-04", "2020-03-05", "2020-03-06", "2020-03-07", "2020-03-08", "2020-03-09", "2020-03-10", "2020-03-11", "2020-03-12", "2020-03-13", "2020-03-14", "2020-03-15", "2020-03-16", "2020-03-17", "2020-03-18", "2020-03-19", "2020-03-20", "2020-03-21", "2020-03-22", "2020-03-23", "2020-03-24", "2020-03-25", "2020-03-26", "2020-03-27", "2020-03-28", "2020-03-29", "2020-03-30", "2020-03-31", "2020-04-01", "2020-04-02", "2020-04-03", "2020-04-04", "2020-04-05", "2020-04-06", "2020-04-07", "2020-04-08", "2020-04-09", "2020-04-10", "2020-04-11", "2020-04-12", "2020-04-13", "2020-04-14", "2020-04-15", "2020-04-16", "2020-04-17", "2020-04-18", "2020-04-19", "2020-04-20", "2020-04-21", "2020-04-22", "2020-04-23", "2020-04-24", "2020-04-25", "2020-04-26", "2020-04-27", "2020-04-28", "2020-04-29", "2020-04-30", "2020-05-01", "2020-05-02", "2020-05-03", "2020-05-04", "2020-05-05", "2020-05-06", "2020-05-07", "2020-05-08", "2020-05-09", "2020-05-10", "2020-05-11", "2020-05-12", "2020-05-13", "2020-05-14", "2020-05-15", "2020-05-16", "2020-05-17", "2020-05-18", "2020-05-19", "2020-05-20", "2020-05-21", "2020-05-22", "2020-05-23", "2020-05-24", "2020-05-25", "2020-05-26", "2020-05-27", "2020-05-28", "2020-05-29", "2020-05-30", "2020-05-31", "2020-06-01", "2020-06-02", "2020-06-03", "2020-06-04"))
    }
    \item{\code{new_deaths}}{c(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 3, 11, 0, 9, 15, 15, 25, 25, 26, 38, 43, 46, 46, 57, 65, 66, 72, 73, 86, 89, 97, 108, 97, 255, 13, 144, 142, 106, 98, 139, 116,  119, 112, 104, 156, 79, 64, 38, 57, 64, 58, 67, 72, 84, 80, 102, 104, 97, 228, 208, 273, 326, 350, 442, 366, 751, 608, 813, 959, 1063, 1379, 1711, 1675, 1847, 2303, 2618, 2807, 3481, 3557, 3191, 3983, 4642, 5040, 4993, 6754, 6274, 4703, 5346, 7632, 6548, 7443, 7018, 5983, 5326, 5329, 7655, 10542, 8551, 8447, 6398,  4951, 5397, 7365, 6041, 7354, 5394, 6141, 3862, 4897, 6569, 6577, 5693, 5571, 4749, 3643, 3995, 5970, 6137, 5668, 5183, 4380, 3359, 3553, 5756, 5113, 5282, 5111, 4202, 2854, 3256, 5185, 4756, 5371, 4641, 3927, 1069, 3355, 3955, 5200, 4643, 4694, 4018, 2865, 3117, 4570, 5562)}
  }
}
\details{x values are in Date format. So, xmax must be as.Date() too.
}
\source{"https://covid.ourworldindata.org/data/owid-covid-data.csv"
}
\references{Roser M, Ritchie H, Ortiz-Ospina E, Hasel J. Coronavirus Pandemic (COVID-19) [Internet]. 2020 [cited 2020 May 15]. Available from: https://ourworldindata.org/coronavirus
}
\examples{
data(worldCOVID)
## maybe str(worldCOVID) ; plot(worldCOVID) ...
}
\keyword{datasets}
